+++ example/cumulus-tf/outputs.tf
@@ -25,7 +25,7 @@ output "report_executions_sns_topic_arn" {
 }
 
 output "report_granules_sns_topic_arn" {
-  value = module.cumulus.report_executions_sns_topic_arn
+  value = module.cumulus.report_granules_sns_topic_arn
 }
 
 output "report_pdrs_sns_topic_arn" {
