+++ packages/api/lambdas/publish-excecutions.js
@@ -6,7 +6,7 @@ const { publishSnsMessage } = require('@cumulus/common/aws');
 const handler = async (event) => {
   const records = event.Records.map(attr.unwrap);
 
-  return Promise.all(
+  await Promise.all(
     records.map((record) => publishSnsMessage(
       process.env.execution_sns_topic_arn,
       record
