+++ packages/api/webpack.config.js
@@ -14,6 +14,7 @@ module.exports = {
     bulkOperation: './lambdas/bulk-operation.js',
     cleanExecutions: './lambdas/cleanExecutions.js',
     createReconciliationReport: './lambdas/create-reconciliation-report.js',
+    cwSfExecutionEventToDb: './lambdas/cw-sf-execution-event-to-db.js',
     dbIndexer: './lambdas/db-indexer.js',
     distribution: './app/distribution.js',
     emsDistributionReport: './lambdas/ems-distribution-report.js',
@@ -24,8 +25,8 @@ module.exports = {
     indexFromDatabase: './lambdas/index-from-database.js',
     messageConsumer: './lambdas/message-consumer.js',
     payloadLogger: './lambdas/payload-logger.js',
+    publishExecutions: './lambdas/publish-executions.js',
     publishReports: './lambdas/publish-reports.js',
-    reportExecutions: './lambdas/report-executions.js',
     reportGranules: './lambdas/report-granules.js',
     reportPdrs: './lambdas/report-pdrs.js',
     sfScheduler: './lambdas/sf-scheduler.js',
