+++ tf-modules/archive/publish_reports.tf
@@ -59,7 +59,6 @@ data "aws_iam_policy_document" "publish_reports_policy_document" {
       "SNS:Publish"
     ]
     resources = [
-      aws_sns_topic.report_executions_topic.arn,
       aws_sns_topic.report_granules_topic.arn,
       aws_sns_topic.report_pdrs_topic.arn
     ]
