+++ tf-modules/workflow/main.tf
@@ -47,6 +47,18 @@ resource "aws_lambda_permission" "publish_reports" {
   source_arn    = aws_cloudwatch_event_rule.state_machine_execution_rule.arn
 }
 
+resource "aws_cloudwatch_event_target" "cw_sf_execution_event_to_db" {
+  rule = aws_cloudwatch_event_rule.state_machine_execution_rule.name
+  arn  = var.workflow_config.cw_sf_execution_event_to_db_lambda_function_arn
+}
+
+resource "aws_lambda_permission" "cw_sf_execution_event_to_db" {
+  action        = "lambda:InvokeFunction"
+  function_name = var.workflow_config.cw_sf_execution_event_to_db_lambda_function_arn
+  principal     = "events.amazonaws.com"
+  source_arn    = aws_cloudwatch_event_rule.state_machine_execution_rule.arn
+}
+
 locals {
   workflow_info = jsonencode({
     name       = var.name
