+++ tf-modules/workflow/variables.tf
@@ -23,6 +23,7 @@ variable "system_bucket" {
 variable "workflow_config" {
   description = "Configuration object with ARNs for workflow integration (Role ARN for executing workflows and Lambda ARNs to trigger on workflow execution)"
   type = object({
+    cw_sf_execution_event_to_db_lambda_function_arn = string
     publish_reports_lambda_function_arn = string
     sf_semaphore_down_lambda_function_arn = string
     state_machine_role_arn = string
